refactor(components): Add ComboboxContent Tests (#1537)
move some Content specific tests into own test file
